moved InitService to tests
